Move RenderedContent to its own dir
