+++ src/inspect_ai/_view/www/src/components/RenderedContent/RenderedContent.mjs
@@ -1,12 +1,12 @@
 import { html } from "htm/preact";
 
-import { ApplicationIcons } from "../appearance/Icons.mjs";
-import { FontSize } from "../appearance/Fonts.mjs";
+import { ApplicationIcons } from "../../appearance/Icons.mjs";
+import { FontSize } from "../../appearance/Fonts.mjs";
 
-import { ANSIDisplay } from "./AnsiDisplay.mjs";
-import { MetaDataView } from "./MetaDataView.mjs";
-import { ChatView } from "./ChatView.mjs";
-import { formatNumber } from "../utils/Format.mjs";
+import { ANSIDisplay } from "../AnsiDisplay.mjs";
+import { MetaDataView } from "../MetaDataView.mjs";
+import { ChatView } from "../ChatView.mjs";
+import { formatNumber } from "../../utils/Format.mjs";
 
 export const RenderedContent = ({
   id,
